SIG-1503 reorganize tests
