+++ api/app/tests/apps/sigmax/stuf_protocol/outgoing/data/example-bv03.xml
