+++ api/app/tests/apps/sigmax/stuf_protocol/outgoing/test_creeerZaak_Lk01.py
@@ -3,20 +3,15 @@ Test suite for Sigmax message generation.
 """
 import copy
 import logging
-import os
 from datetime import timedelta
-from unittest import mock
 
 from django.contrib.gis.geos import Point
-from django.test import TestCase, override_settings
+from django.test import TestCase
 from django.utils import timezone
 from lxml import etree
 from xmlunittest import XmlTestMixin
 
-from signals.apps.sigmax.models import CityControlRoundtrip
-from signals.apps.sigmax.stuf_protocol import outgoing
-from signals.apps.sigmax.stuf_protocol.exceptions import SigmaxException
-from signals.apps.sigmax.stuf_protocol.outgoing import _generate_sequence_number, handle
+from signals.apps.sigmax.stuf_protocol.outgoing import _generate_sequence_number
 from signals.apps.sigmax.stuf_protocol.outgoing.creeerZaak_Lk01 import (
     SIGMAX_REQUIRED_ADDRESS_FIELDS,
     SIGMAX_STADSDEEL_MAPPING,
@@ -24,59 +19,14 @@ from signals.apps.sigmax.stuf_protocol.outgoing.creeerZaak_Lk01 import (
     _generate_creeerZaak_Lk01,
     _generate_omschrijving
 )
-from signals.apps.sigmax.stuf_protocol.outgoing.voegZaakdocumentToe_Lk01 import (
-    _generate_voegZaakdocumentToe_Lk01
-)
 from signals.apps.signals.models import Priority, Signal
 from tests.apps.signals.factories import SignalFactory, SignalFactoryValidLocation
 from tests.apps.signals.valid_locations import STADHUIS
 
-LOG_FORMAT = '%(asctime)-15s - %(name)s - %(message)s'
-logging.basicConfig(format=LOG_FORMAT, level=logging.DEBUG)
-logging.disable(logging.NOTSET)
 logger = logging.getLogger(__name__)
-
-REQUIRED_ENV = {'SIGMAX_AUTH_TOKEN': 'TEST', 'SIGMAX_SERVER': 'https://example.com'}
-DATA_DIR = os.path.join(
-    os.path.dirname(os.path.abspath(__file__)),
-    'data'
-)
-
-
-class TestHandle(TestCase):
-    def setUp(self):
-        self.signal = SignalFactory.create()
-
-    @mock.patch('signals.apps.sigmax.stuf_protocol.outgoing.MAX_ROUND_TRIPS', 2)
-    @mock.patch('signals.apps.sigmax.stuf_protocol.outgoing.send_creeerZaak_Lk01', autospec=True)
-    @mock.patch(
-        'signals.apps.sigmax.stuf_protocol.outgoing.send_voegZaakdocumentToe_Lk01', autospec=True)
-    def test_too_many(self,
-                      patched_send_voegZaakdocumentToe_Lk01,
-                      patched_send_creeerZaak_Lk01,):
-        CityControlRoundtrip.objects.create(_signal=self.signal)
-        CityControlRoundtrip.objects.create(_signal=self.signal)
-        CityControlRoundtrip.objects.create(_signal=self.signal)
-
-        with self.assertRaises(SigmaxException):
-            handle(self.signal)
-        patched_send_voegZaakdocumentToe_Lk01.assert_not_called()
-
-    @mock.patch('signals.apps.sigmax.stuf_protocol.outgoing.send_creeerZaak_Lk01', autospec=True)
-    @mock.patch(
-        'signals.apps.sigmax.stuf_protocol.outgoing.send_voegZaakdocumentToe_Lk01', autospec=True)
-    def test_success_message(self,
-                             patched_send_voegZaakdocumentToe_Lk01,
-                             patched_send_creeerZaak_Lk01):
-        success_message = handle(self.signal)
-        self.assertIn(
-            '{}.01'.format(self.signal.sia_id),
-            success_message
-        )
 
 
 class TestOutgoing(TestCase, XmlTestMixin):
-
     def test_generate_creeerZaak_Lk01(self):
         current_tz = timezone.get_current_timezone()
         location = Point(4.1234, 52.1234)
@@ -138,24 +88,6 @@ class TestOutgoing(TestCase, XmlTestMixin):
             '<ZKN:einddatumGepland>{}</ZKN:einddatumGepland>'.format(
                 incident_date_end.astimezone(current_tz).strftime('%Y%m%d')),
             xml_message)
-
-    def test_generate_voegZaakdocumentToe_Lk01(self):
-        signal = SignalFactoryValidLocation.create()
-        seq_no = _generate_sequence_number(signal)
-        xml_message = _generate_voegZaakdocumentToe_Lk01(signal, seq_no)
-        self.assertXmlDocument(xml_message)
-
-        self.assertIn(
-            f'<ZKN:identificatie>{signal.sia_id}.{seq_no}</ZKN:identificatie>',
-            xml_message
-        )
-
-    def test_generate_voegZaakdocumentToe_Lk01_escaping(self):
-        poison = SignalFactoryValidLocation.create()
-        poison.text = '<poison>tastes nice</poison>'
-        seq_no = '02'
-        xml_message = _generate_voegZaakdocumentToe_Lk01(poison, seq_no)
-        self.assertTrue('<poison>' not in xml_message)
 
 
 class TestGenerateOmschrijving(TestCase):
@@ -242,24 +174,6 @@ class TestAddressMatchesSigmaxExpectation(TestCase):
         # simulate empty address field
         self.assertEqual(False, _address_matches_sigmax_expectation(None))
         self.assertEqual(False, _address_matches_sigmax_expectation({}))
-
-
-class TestGenerateSequenceNumber(TestCase):
-    def setUp(self):
-        self.signal = SignalFactory.create()
-
-    def test_initial_sequence_number(self):
-        self.assertEqual(CityControlRoundtrip.objects.count(), 0)
-        self.assertEqual(_generate_sequence_number(self.signal), '01')
-
-    def test_fourth_sequence_number(self):
-        self.assertEqual(CityControlRoundtrip.objects.count(), 0)
-        CityControlRoundtrip.objects.bulk_create([
-            CityControlRoundtrip(_signal=self.signal),
-            CityControlRoundtrip(_signal=self.signal),
-            CityControlRoundtrip(_signal=self.signal),
-        ])
-        self.assertEqual(_generate_sequence_number(self.signal), '04')
 
 
 class TestGenerateCreeerZaakLk01Message(TestCase, XmlTestMixin):
@@ -350,99 +264,3 @@ class TestGenerateCreeerZaakLk01Message(TestCase, XmlTestMixin):
         namespaces = {'zaak': 'http://www.egem.nl/StUF/sector/zkn/0310'}
         found = root.xpath('//zaak:object//zaak:heeftBetrekkingOp', namespaces=namespaces)
         self.assertEqual(found, [])
-
-
-class TestVoegZaakDocumentToeLk01Message(TestCase):
-
-    def setUp(self):
-        self.signal: Signal = SignalFactoryValidLocation.create()
-        self.seq_no = _generate_sequence_number(self.signal)
-
-    def test_is_xml(self):
-        signal = self.signal
-        xml = _generate_voegZaakdocumentToe_Lk01(signal, self.seq_no)
-        try:
-            etree.fromstring(xml)
-        except Exception:
-            self.fail('Cannot parse STUF message as XML')
-
-    def test_escaping(self):
-        poison: Signal = self.signal
-        poison.text = '<poison>tastes nice</poison>'
-        xml = _generate_voegZaakdocumentToe_Lk01(poison, self.seq_no)
-        self.assertTrue('<poison>' not in xml)
-
-
-class TestSendStufMessage(TestCase):
-    @override_settings(
-        SIGMAX_AUTH_TOKEN='',
-        SIGMAX_SERVER='',
-    )
-    def test_no_environment_variables(self):
-        # Check that missing enviroment variables for server and auth token
-        # raises an error when a message is sent.
-        env_override = {'SIGMAX_AUTH_TOKEN': '', 'SIGMAX_SERVER': ''}
-
-        with mock.patch.dict('os.environ', env_override):
-            with self.assertRaises(outgoing.SigmaxException):
-                action = 'http://www.egem.nl/StUF/sector/zkn/0310/CreeerZaak_Lk01'
-                outgoing.stuf._send_stuf_message('TEST BERICHT', action)
-
-    @override_settings(
-        SIGMAX_AUTH_TOKEN=REQUIRED_ENV['SIGMAX_AUTH_TOKEN'],
-        SIGMAX_SERVER=REQUIRED_ENV['SIGMAX_SERVER'],
-    )
-    @mock.patch('signals.apps.sigmax.stuf_protocol.outgoing.stuf._stuf_response_ok', autospec=True)
-    @mock.patch('requests.post', autospec=True)
-    def test_send_message(self, mocked_request_post, mocked_stuf_response_ok):
-        mocked_request_post.return_value.status_code = 200
-        mocked_request_post.return_value.text = 'Message from Sigmax'
-        mocked_stuf_response_ok.return_value = True
-
-        message = 'TEST BERICHT'
-        action = 'http://www.egem.nl/StUF/sector/zkn/0310/CreeerZaak_Lk01'
-        outgoing.stuf._send_stuf_message(message, action)
-
-        # Check that headers are set correctly when sending an STUF message.
-        args, kwargs = mocked_request_post.call_args
-
-        self.assertEqual(mocked_request_post.called, 1)
-        self.assertEqual(kwargs['url'], REQUIRED_ENV['SIGMAX_SERVER'])
-        self.assertEqual(
-            kwargs['headers']['Authorization'],
-            'Basic ' + REQUIRED_ENV['SIGMAX_AUTH_TOKEN']
-        )
-        self.assertEqual(
-            kwargs['headers']['SOAPAction'],
-            'http://www.egem.nl/StUF/sector/zkn/0310/CreeerZaak_Lk01'
-        )
-        self.assertEqual(
-            kwargs['headers']['Content-Type'],
-            'text/xml; charset=UTF-8'
-        )
-        self.assertEqual(
-            b'%d' % len(message),
-            kwargs['headers']['Content-Length']
-        )
-
-
-class TestStufResponseOk(TestCase):
-    def test_Bv03(self):
-        test_xml_file = os.path.join(DATA_DIR, 'example-bv03.xml')
-        with open(test_xml_file, 'rt', encoding='utf-8') as f:
-            test_xml = f.read()
-
-        fake_response = mock.MagicMock()
-        fake_response.text = test_xml
-
-        self.assertEqual(outgoing.stuf._stuf_response_ok(fake_response), True)
-
-    def test_Fo03(self):
-        test_xml_file = os.path.join(DATA_DIR, 'example-fo03.xml')
-        with open(test_xml_file, 'rt', encoding='utf-8') as f:
-            test_xml = f.read()
-
-        fake_response = mock.MagicMock()
-        fake_response.text = test_xml
-
-        self.assertEqual(outgoing.stuf._stuf_response_ok(fake_response), False)
