+++ api/app/tests/apps/sigmax/stuf_protocol/outgoing/test_pdf.py
